Remove rest of print functionality, move adapter into current fulder
