+++ src/app/psat/psat-report/psat-report.component.html
@@ -52,19 +52,6 @@
     </div>
     <!-- End .report-header -->
 
-    <div *ngIf="showPrintDiv && !inRollup" class="hide-print print-div row justify-content-center">
-        <div class="col-12 print-col">
-            <h4>Building Print View</h4>
-            <h5>One Moment Please</h5>
-            <span class="fa fa-print"></span>
-            <div class="spinner">
-                <div class="bounce1"></div>
-                <div class="bounce2"></div>
-                <div class="bounce3"></div>
-            </div>
-        </div>
-    </div>
-
     <div class="report-data" *ngIf="assessment.psat.setupDone && assessment.psat.valid.isValid">
         <div class="d-none d-lg-flex">
             <ul #reportBtns class="tabs primary hide-print flex-fill">
@@ -129,30 +116,29 @@
             </app-export-button>
         </div>
 
-        <div class="results scroll-item print-height" *ngIf="currentTab == 'facilityInfo' || showPrintView"
+        <div class="results scroll-item print-height" *ngIf="currentTab == 'facilityInfo'"
             [ngStyle]="{'height.px':reportContainerHeight}">
             <app-facility-info-summary [assessment]="assessment"></app-facility-info-summary>
         </div>
         <div class="input-data results scroll-item print-height"
-            *ngIf="(currentTab == 'inputData' && !showPrintView) || (showPrintView && printOptions.printInputData)"
+            *ngIf="currentTab == 'inputData'"
             [ngStyle]="{'height.px':reportContainerHeight}">
             <app-input-summary [psat]="assessment.psat" [settings]="settings">
             </app-input-summary>
         </div>
         <div class="results scroll-item print-height"
-            *ngIf="(currentTab == 'results' && !showPrintView) || (showPrintView && printOptions.printResults)"
+            *ngIf="currentTab == 'results'"
             [ngStyle]="{'height.px':reportContainerHeight}">
             <app-output-summary [settings]="settings" [inRollup]="inRollup" [assessment]="assessment">
             </app-output-summary>
         </div>
         <div class="results scroll-item print-height"
-            *ngIf="(currentTab == 'reportGraphs' && !showPrintView) || (showPrintView && (printOptions.printReportGraphs || printOptions.printReportSankey))"
+            *ngIf="currentTab == 'reportGraphs'"
             [ngStyle]="{'height.px':reportContainerHeight}">
-            <app-psat-report-graphs [settings]="settings" [psat]="assessment.psat" [printView]="showPrintView"
-                [printSankey]="printOptions.printReportSankey" [printGraphs]="printOptions.printReportGraphs">
+            <app-psat-report-graphs [settings]="settings" [psat]="assessment.psat">
             </app-psat-report-graphs>
         </div>
-        <div class="results scroll-item print-height" *ngIf="currentTab == 'reportSankey' && !showPrintView"
+        <div class="results scroll-item print-height" *ngIf="currentTab == 'reportSankey'"
             [ngStyle]="{'height.px':reportContainerHeight}">
             <app-psat-report-sankey [settings]="settings" [assessment]="assessment"></app-psat-report-sankey>
         </div>
@@ -165,6 +151,3 @@
             see the resulting data.</p>
     </div>
 </div>
-
-
-<app-print-options-menu *ngIf="showPrintMenu"></app-print-options-menu>
